requirements.txt
fastapi
tortoise-orm
uvicorn
pyjwt
passlib[bcrypt]
python-multipart
bcrypt
pytest
python-dotenv